Removed templates from abstraction
